'input_pages 분할'
