+++ Sangsoo/study_8th/lib/input_page.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+
+class InputPage extends StatefulWidget {
+  @override
+  _InputPageState createState() => _InputPageState();
+}
+
+class _InputPageState extends State<InputPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        centerTitle: true,
+        title: Text('BMI CALCULATOR'),
+      ),
+      body: Container(
+        color: Color(0xFF1D1E33),
+        height: 200.0,
+        width:170.0
+      ),
+    );
+  }
+}
